+++ pytorch_lightning/accelerators/base_accelerator.py
@@ -27,6 +27,9 @@ class Accelerator(object):
         self.trainer = trainer
         self.cluster_environment = cluster_environment
         self.dist = AttributeDict(rank=0, device=None)
+        self.train_loop = self.trainer.train
+        self.validation_loop = self.trainer.run_evaluation
+        self.test_loop = self.trainer.run_evaluation
 
     def setup(self, model):
         pass
