+++ pytorch_lightning/accelerators/gpu_backend.py
@@ -14,7 +14,7 @@
 
 import torch
 
-from pytorch_lightning.accelerators.base_backend import Accelerator
+from pytorch_lightning.accelerators.base_accelerator import Accelerator
 from pytorch_lightning.utilities import AMPType
 from pytorch_lightning.distributed.dist import LightningDistributed
