+++ pytorch_lightning/tuner/tuning.py
@@ -29,6 +29,31 @@ class Tuner:
         self.trainer.auto_lr_find = auto_lr_find
         self.trainer.auto_scale_batch_size = auto_scale_batch_size
 
+    def tune(self, model, train_dataloader, val_dataloaders, datamodule):
+        # setup data, etc...
+        self.trainer.train_loop.setup_fit(model, train_dataloader, val_dataloaders, datamodule)
+
+        # hook
+        self.trainer.data_connector.prepare_data(model)
+
+        # Run auto batch size scaling
+        if self.trainer.auto_scale_batch_size:
+            if isinstance(self.trainer.auto_scale_batch_size, bool):
+                self.trainer.auto_scale_batch_size = 'power'
+            self.scale_batch_size(
+                model,
+                mode=self.trainer.auto_scale_batch_size,
+                train_dataloader=train_dataloader,
+                val_dataloaders=val_dataloaders,
+                datamodule=datamodule,
+            )
+            model.logger = self.trainer.logger  # reset logger binding
+
+        # Run learning rate finder:
+        if self.trainer.auto_lr_find:
+            self.internal_find_lr(self.trainer, model)
+            model.logger = self.trainer.logger  # reset logger binding
+
     def scale_batch_size(self,
                          model,
                          mode: str = 'power',
